--- a/core/src/main/java/io/questdb/cairo/map/FixedSizeMapRecord.java
+++ b/core/src/main/java/io/questdb/cairo/map/FixedSizeMapRecord.java
@@ -34,29 +34,34 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 /**
- * Provides Record access interface for FastMap key-value pairs with fixed-size keys.
+ * Provides Record access interface for FixedSizeMap key-value pairs.
  * <p>
  * Uses an offsets array to speed up key and value column look-ups.
  */
-final class Unordered16MapRecord implements MapRecord {
+final class FixedSizeMapRecord implements MapRecord {
     private final long[] columnOffsets;
     private final Long256Impl[] keyLong256A;
     private final Long256Impl[] keyLong256B;
-    private final Unordered16MapValue value;
+    private final long keySize;
+    private final FixedSizeMapValue value;
     private final long[] valueOffsets;
     private final long valueSize;
+    private long keyAddress;
     private long limit;
-    private long startAddress;
     private IntList symbolTableIndex;
     private RecordCursor symbolTableResolver;
+    private long valueAddress;
 
-    Unordered16MapRecord(
+    FixedSizeMapRecord(
+            long keySize,
             long valueSize,
             long[] valueOffsets,
-            Unordered16MapValue value,
+            FixedSizeMapValue value,
             @NotNull @Transient ColumnTypes keyTypes,
             @Nullable @Transient ColumnTypes valueTypes
     ) {
+        assert keySize >= 0;
+        this.keySize = keySize;
         this.valueSize = valueSize;
         this.valueOffsets = valueOffsets;
         this.value = value;
@@ -76,7 +81,7 @@ final class Unordered16MapRecord implements MapRecord {
 
         Long256Impl[] long256A = null;
         Long256Impl[] long256B = null;
-        long offset = 0;
+        int offset = 0;
         for (int i = 0, n = keyTypes.getColumnCount(); i < n; i++) {
             final int columnType = keyTypes.getColumnType(i);
             if (ColumnType.tagOf(columnType) == ColumnType.LONG256) {
@@ -95,8 +100,6 @@ final class Unordered16MapRecord implements MapRecord {
             offset += size;
         }
 
-        assert offset <= Unordered16Map.KEY_SIZE;
-        offset = Unordered16Map.KEY_SIZE;
         if (valueTypes != null) {
             for (int i = 0, n = valueTypes.getColumnCount(); i < n; i++) {
                 int columnType = valueTypes.getColumnType(i);
@@ -121,24 +124,26 @@ final class Unordered16MapRecord implements MapRecord {
         this.keyLong256B = long256B;
     }
 
-    private Unordered16MapRecord(
+    private FixedSizeMapRecord(
+            long keySize,
             long valueSize,
             long[] valueOffsets,
             long[] columnOffsets,
             Long256Impl[] keyLong256A,
             Long256Impl[] keyLong256B
     ) {
+        this.keySize = keySize;
         this.valueSize = valueSize;
         this.valueOffsets = valueOffsets;
         this.columnOffsets = columnOffsets;
-        this.value = new Unordered16MapValue(valueSize, valueOffsets);
+        this.value = new FixedSizeMapValue(valueSize, valueOffsets);
         this.keyLong256A = keyLong256A;
         this.keyLong256B = keyLong256B;
     }
 
     @SuppressWarnings("MethodDoesntCallSuperMethod")
     @Override
-    public Unordered16MapRecord clone() {
+    public FixedSizeMapRecord clone() {
         final Long256Impl[] long256A;
         final Long256Impl[] long256B;
 
@@ -157,19 +162,19 @@ final class Unordered16MapRecord implements MapRecord {
             long256A = null;
             long256B = null;
         }
-        return new Unordered16MapRecord(valueSize, valueOffsets, columnOffsets, long256A, long256B);
+        return new FixedSizeMapRecord(keySize, valueSize, valueOffsets, columnOffsets, long256A, long256B);
     }
 
     @Override
     public void copyToKey(MapKey destKey) {
-        Unordered16Map.Key destBaseKey = (Unordered16Map.Key) destKey;
-        destBaseKey.copyFromRawKey(startAddress);
+        FixedSizeMap.Key destFastKey = (FixedSizeMap.Key) destKey;
+        destFastKey.copyFromRawKey(keyAddress, keySize);
     }
 
     @Override
     public void copyValue(MapValue destValue) {
-        Unordered16MapValue destFastValue = (Unordered16MapValue) destValue;
-        destFastValue.copyRawValue(startAddress + Unordered16Map.KEY_SIZE);
+        FixedSizeMapValue destFastValue = (FixedSizeMapValue) destValue;
+        destFastValue.copyRawValue(valueAddress);
     }
 
     @Override
@@ -264,9 +269,9 @@ final class Unordered16MapRecord implements MapRecord {
 
     @Override
     public long getRowId() {
-        // Important invariant: we assume that the map doesn't grow after the first getRowId() call.
+        // Important invariant: we assume that the FastMap doesn't grow after the first getRowId() call.
         // Otherwise, row ids returned by this method may no longer point at a valid memory address.
-        return startAddress;
+        return keyAddress;
     }
 
     @Override
@@ -286,18 +291,17 @@ final class Unordered16MapRecord implements MapRecord {
 
     @Override
     public MapValue getValue() {
-        return value.of(startAddress, limit, false);
+        return value.of(keyAddress, valueAddress, limit, false);
     }
 
     @Override
     public long keyHashCode() {
-        long key1 = Unsafe.getUnsafe().getLong(startAddress);
-        long key2 = Unsafe.getUnsafe().getLong(startAddress + 8L);
-        return Hash.hashLong128_64(key1, key2);
+        return Hash.hashMem64(keyAddress, keySize);
     }
 
     public void of(long address) {
-        this.startAddress = address;
+        this.keyAddress = address;
+        this.valueAddress = address + keySize;
     }
 
     public void setLimit(long limit) {
@@ -311,14 +315,13 @@ final class Unordered16MapRecord implements MapRecord {
     }
 
     private long addressOfColumn(int index) {
-        return startAddress + columnOffsets[index];
+        return keyAddress + columnOffsets[index];
     }
 
     @NotNull
     private Long256 getLong256Generic(Long256Impl[] keyLong256, int columnIndex) {
-        long address = addressOfColumn(columnIndex);
         Long256Impl long256 = keyLong256[columnIndex];
-        long256.fromAddress(address);
+        long256.fromAddress(addressOfColumn(columnIndex));
         return long256;
     }
 }