Fixes the following performance issue observed on the demo:
```sql
-- 1.9s
SELECT timestamp, symbol, side, min(price)
FROM trades
WHERE timestamp in '2024-01'
SAMPLE BY 30s;

-- 1.3s
SELECT timestamp, symbol, side, min(price), max(price), avg(price)
FROM trades
WHERE timestamp in '2024-01'
SAMPLE BY 30s;
```

Includes the following:
* Removes `Unordered2Map` and `Unordered16Map` as they don't lead to better performance in all cases.
* Splits `OrderedMap` into `VarSizeMap` and `FixedSizeMap`. The latter doesn't cache hash codes, so its overhead per entry is 4 bytes while for `VarSizeMap` it's 8 bytes.
* Changes `MapFactory` to use `Unordered4Map` and `Unordered8Map` only for keys of exactly 4 and 8 bytes.