@ideoma 2024-06-19
#### [PR Coverage check]

:heart_eyes: **pass** : 526 / 562 (93.59%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/cairo/map/VarSizeMapValue.java|2|3|66.67%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered8Map.java|30|39|76.92%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered4Map.java|32|38|84.21%|
|:large_blue_circle:|io/questdb/cairo/map/FixedSizeMapValue.java|10|11|90.91%|
|:large_blue_circle:|io/questdb/cairo/map/FixedSizeMap.java|301|317|94.95%|
|:large_blue_circle:|io/questdb/cairo/map/VarSizeMap.java|84|87|96.55%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/DistinctTimeSeriesRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/PropServerConfiguration.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/VarSizeMapCursor.java|7|7|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/VarSizeMapRecord.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/FixedSizeMapRecord.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/cairo/DefaultCairoConfiguration.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/JoinRecordMetadata.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/PropertyKey.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/cairo/CairoConfigurationWrapper.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/FixedSizeMapCursor.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/MapFactory.java|33|33|100.00%|