@puzpuzpuz 2024-06-25
Closing this PR. Will submit a more light-weight version of it